Rename original anchor contract to anchor-lend
