+++ contracts/andromeda-anchor-lend/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "andromeda-anchor"
+name = "andromeda-anchor-lend"
 version = "0.1.0"
 edition = "2018"
