+++ contracts/andromeda-anchor-lend/src/testing/tests.rs
@@ -17,21 +17,18 @@ use crate::testing::mock_querier::{
 };
 use ado_base::ADOContract;
 use anchor_token::gov::{Cw20HookMsg as GovCw20HookMsg, ExecuteMsg as GovExecuteMsg};
-use andromeda_protocol::anchor::{
+use andromeda_protocol::anchor_lend::{
     BLunaHubCw20HookMsg, BLunaHubExecuteMsg, Cw20HookMsg, ExecuteMsg, InstantiateMsg,
 };
 use common::{
-    ado_base::{
-        recipient::{Recipient},
-        AndromedaMsg,
-    },
+    ado_base::{recipient::Recipient, AndromedaMsg},
     error::ContractError,
 };
 use cw20::{Cw20ExecuteMsg, Cw20ReceiveMsg};
 use cw_asset::AssetInfo;
 use moneymarket::{
     custody::{Cw20HookMsg as CustodyCw20HookMsg, ExecuteMsg as CustodyExecuteMsg},
-    market::{ExecuteMsg as MarketExecuteMsg},
+    market::ExecuteMsg as MarketExecuteMsg,
     overseer::ExecuteMsg as OverseerExecuteMsg,
 };
