+++ packages/andromeda_protocol/src/lib.rs
@@ -1,7 +1,7 @@
 pub mod address_list;
 pub mod airdrop;
-pub mod anchor;
 pub mod anchor_earn;
+pub mod anchor_lend;
 pub mod astroport;
 pub mod auction;
 pub mod common;
